Fixed the create goal and deadline bug
